+++ src/__tests__/actions/app/BalanceActions.test.js
@@ -33,6 +33,12 @@ describe('Test Balance Actions', () => {
                 'access-control-allow-credentials': 'true'
             })
             .get(apiResourceUrl)
+            .query({
+                page: 0,
+                perPage: 5,
+                orderBy: "id",
+                orderWay: "asc"
+            })
             .reply(200, dateGetReturn);
         store.dispatch(balanceActions.index());
         expect(store.getActions()).toContainEqual({
@@ -57,6 +63,12 @@ describe('Test Balance Actions', () => {
                 'access-control-allow-credentials': 'true'
             })
             .get(apiResourceUrl)
+            .query({
+                page: 0,
+                perPage: 5,
+                orderBy: "id",
+                orderWay: "asc"
+            })
             .reply(401);
         store.dispatch(balanceActions.index());
         expect(store.getActions()).toContainEqual({