+++ src/api/app/BankAccountTransactions.js
@@ -1,0 +1,12 @@
+import ApiCrud from "../base/ApiCrud";
+
+
+export default class BankAccountTransactions extends ApiCrud {
+
+    getBaseUrl() {
+        return "/bank-account-transactions";
+    }
+
+}
+
+export const instance = new BankAccountTransactions();