+++ src/views/Router.js
@@ -1,7 +1,8 @@
 import { Routes, Route } from "react-router-dom";
 import HomePage from "./HomePage/HomePage";
 import DashboardPage from "./DashboardPage/DashboardPage";
-import { PATH_DASHBOARD, PATH_HOME } from "../utils/constants/constant";
+import BankAccountTransactionsPage from "./BankAccountTransactionsPage/BankAccountTransactionsPage";
+import { PATH_BANK_ACCOUNT_TRANSACTIONS, PATH_DASHBOARD, PATH_HOME } from "../utils/constants/constant";
 import AuthRouter from "./AuthRouter";
 
 export const Router = () => {
@@ -9,6 +10,7 @@ export const Router = () => {
         <Routes>
             <Route path={PATH_HOME} element={<AuthRouter type="guest"><HomePage /></AuthRouter>} />
             <Route path={PATH_DASHBOARD} element={<AuthRouter type="private"><DashboardPage /></AuthRouter>} />
+            <Route path={PATH_BANK_ACCOUNT_TRANSACTIONS} element={<AuthRouter type="private"><BankAccountTransactionsPage /></AuthRouter>} />
         </Routes>
     );
 }