+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-# Next version
+# 0.3.0
++ Fix Semaphore
++ Upgrade dependencies
++ Add `bank-account-transactions` api, reducer, page
++ Check for `content` in items
++ Add Bank Account Transaction Page
++ Add Remote Pagination
++ Fetch APP version from package.json
++ Add remote sort by
+
+## 0.2.0
 + Add CircleCI
 + Add balance api and reducer
 + Add KPI